Add placement type customization support
