+++ frontend/src/employee-frontend/components/common/CareTypeLabel.tsx
@@ -8,7 +8,7 @@ import styled from 'styled-components'
 import { useTranslation } from '../../state/i18n'
 import { CareTypeLabel } from '../../types'
 import colors from 'lib-customizations/common'
-import { PlacementType } from 'lib-common/api-types/serviceNeed/common'
+import { PlacementType } from 'lib-customizations/types'
 
 export interface Props {
   type?: CareTypeLabel
