+++ frontend/src/employee-mobile-frontend/components/attendances/child-info/ChildSensitiveInfo.tsx
@@ -18,7 +18,7 @@ import { faPhone } from 'lib-icons'
 
 import { Child } from '../../../api/attendances'
 import { useTranslation } from '../../../state/i18n'
-import { PlacementType } from '../../../types'
+import { PlacementType } from 'lib-customizations/types'
 
 interface Props {
   child: Child | null
